backend/pyproject.toml
[project]
name = "t3-chat-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pypdf2>=3.0.1",
    "sse-starlette>=2.3.6",
    "supabase>=2.15.3",
    "uvicorn>=0.34.3",
]
